remove more empty impls and all ops trait impls to separate module
