+++ polars/polars-core/src/series/implementations/list.rs
@@ -339,25 +339,6 @@ impl SeriesTrait for SeriesWrap<ListChunked> {
             .map(|ca| ca.into_series())
     }
 
-    fn pow(&self, exponent: f64) -> Result<Series> {
-        let f_err = || {
-            Err(PolarsError::InvalidOperation(
-                format!("power operation not supported on dtype {:?}", self.dtype()).into(),
-            ))
-        };
-
-        match self.dtype() {
-            DataType::Utf8 | DataType::List(_) | DataType::Boolean => f_err(),
-            DataType::Float32 => Ok(self.0.pow_f32(exponent as f32).into_series()),
-            _ => Ok(self.0.pow_f64(exponent).into_series()),
-        }
-    }
-
-    #[cfg(feature = "is_first")]
-    fn is_first(&self) -> Result<BooleanChunked> {
-        self.0.is_first()
-    }
-
     #[cfg(feature = "object")]
     fn as_any(&self) -> &dyn Any {
         &self.0
